Merge two sorted lists
Two friends were playing a game of cards with only a single colour (i.e Diamonds). One of them was having a sorted list of all six cards and the other one was having a sorted list of the other seven cards from the deck of thirteen cards in diamond. Write a code that uses the two set of sorted cards to finally get a single set of sorted cards.

Input:
6 
1 4 8 11 12 13 
6 
2 3 5 6 7 9
Output:
1 0 0 0 0 0 0 0 0 0 0 0 
1 2 0 0 0 0 0 0 0 0 0 0 
1 2 3 0 0 0 0 0 0 0 0 0 
1 2 3 4 0 0 0 0 0 0 0 0 
1 2 3 4 5 0 0 0 0 0 0 0 
1 2 3 4 5 6 0 0 0 0 0 0 
1 2 3 4 5 6 7 0 0 0 0 0 
1 2 3 4 5 6 7 8 0 0 0 0 
1 2 3 4 5 6 7 8 9 0 0 0 
1 2 3 4 5 6 7 8 9 11 0 0 
1 2 3 4 5 6 7 8 9 11 12 0 
1 2 3 4 5 6 7 8 9 11 12 13