Sorts the given list in ascending order
You are given some integers and have been asked to sort them in a particular fashion. In this approach, every time you choose an integer from the given list and place it at a position you make sure that all the numbers left to it are either less than or equal to that number and all the elements right of it are greater than that number. Proceed in this way and finally write a code that sorts the given list in ascending order.